Added dictionary showing window - dictionary.fxml and DictionaryController. Added needed function for sorting dictionary to Indexer. Fixed and updated Parse. Added necessary functions and fixes for controller to display dictionary.

+++ src/Controller/Controller.java
@@ -4,11 +4,16 @@ import Part_1.Indexer;
 import Part_1.Parse;
 import Part_1.ReadFile;
 import javafx.event.ActionEvent;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.TextField;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.CheckBox;
 import javafx.stage.DirectoryChooser;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
 import java.io.File;
 
 /**
@@ -21,17 +26,16 @@ public class Controller {
     public CheckBox stemmingCheckBox;
     public ChoiceBox<String> languageChoiceBox;
     public static String postingPathText;
+    private boolean startsIndexing = false;
+    private boolean alreadyIndexed = false;
 
     /**
      * opens a Directory Chooser window in order to choose a directory path for the corpus and for the stop words file
      * @param actionEvent - unused
      */
     public void onCorpusBrowse(ActionEvent actionEvent) {
-        DirectoryChooser directoryChooser = new DirectoryChooser();
-        directoryChooser.setTitle("Choose a path");
-        File selectedDirectory = directoryChooser.showDialog(corpusPath.getScene().getWindow());
-        if (selectedDirectory != null)
-            corpusPath.setText(selectedDirectory.getAbsolutePath());
+        if (!startsIndexing)
+            chooseAndSaveDirectoryPath(corpusPath);
     }
 
     /**
@@ -39,11 +43,8 @@ public class Controller {
      * @param actionEvent - unused
      */
     public void onPostingBrowse(ActionEvent actionEvent) {
-        DirectoryChooser directoryChooser = new DirectoryChooser();
-        directoryChooser.setTitle("Choose a path");
-        File selectedDirectory = directoryChooser.showDialog(postingPath.getScene().getWindow());
-        if (selectedDirectory != null)
-            postingPath.setText(selectedDirectory.getAbsolutePath());
+        if (!startsIndexing)
+            chooseAndSaveDirectoryPath(postingPath);
     }
 
     /**
@@ -51,62 +52,65 @@ public class Controller {
      * @param actionEvent - unused
      */
     public void onActivate(ActionEvent actionEvent) {
-        Parse.stemming = stemmingCheckBox.isSelected();
-        if (postingPath.getText().equals("") || corpusPath.getText().equals("")) {
-            Alert alert = new Alert(Alert.AlertType.ERROR);
-            alert.setContentText("You must fill all necessary paths!");
-            alert.show();
-        }
-        else {
-            if ((new File(postingPath.getText())).exists()) {
-                postingPathText = postingPath.getText();
-                String dirPath = corpusPath.getText();
-                File stopWords = new File(dirPath + "\\stop words");
-                if (stopWords.exists()) {
-                    Parse.stemming = stemmingCheckBox.isSelected();
-                    if (Parse.stemming)
-                        (new File(postingPathText + "\\postingFilesWithStemming")).mkdir();
-                    else
-                        (new File(postingPathText + "\\postingFilesWithoutStemming")).mkdir();
-                    Parse parse = new Parse(dirPath + "\\stop words");
-                    ReadFile readFile = new ReadFile(dirPath);
-                    Indexer indexer = new Indexer();
-                    Thread readFileThread = new Thread(readFile);
-                    Thread parseThread = new Thread(parse);
-                    Thread indexThread = new Thread(indexer);
-                    long startTime = System.nanoTime();
-                    parseThread.start();
-                    readFileThread.start();
-                    indexThread.start();
-                    try {
-                        readFileThread.join();
-                        parseThread.join();
-                        indexThread.join();
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-                    // ------ THE FINAL ALERT BOX INDICATING THE INDEXING PROCESS IS DONE ------
-                    finally {
-                        double totalTimeInSeconds = (System.nanoTime() - startTime)*Math.pow(10,-9);
-                        int totalTimeInMinutes = (int)(totalTimeInSeconds / 60);
-                        int remainingSeconds = (int)(totalTimeInSeconds % 60);
-                        String totalTime = "Total time: " + totalTimeInMinutes + " minutes and " + remainingSeconds + " seconds.";
-                        String docCount = "Total documents indexed: " + ReadFile.docCount;
-                        String termCount = "Total unique words found: " + Indexer.totalUniqueTerms;
-                        Alert alert = new Alert(Alert.AlertType.INFORMATION);
-                        alert.setHeaderText("Indexing Done!");
-                        alert.setContentText(totalTime + "\n" + docCount + "\n" + termCount);
+        if (!alreadyIndexed && !startsIndexing) {
+            Parse.stemming = stemmingCheckBox.isSelected();
+            if (postingPath.getText().equals("") || corpusPath.getText().equals("")) {
+                Alert alert = new Alert(Alert.AlertType.ERROR);
+                alert.setContentText("You must fill all necessary paths!");
+                alert.show();
+            } else {
+                if ((new File(postingPath.getText())).exists()) {
+                    postingPathText = postingPath.getText();
+                    String dirPath = corpusPath.getText();
+                    File stopWords = new File(dirPath + "\\stop words");
+                    if (stopWords.exists()) {
+                        Parse.stemming = stemmingCheckBox.isSelected();
+                        if (Parse.stemming)
+                            (new File(postingPathText + "\\postingFilesWithStemming")).mkdir();
+                        else
+                            (new File(postingPathText + "\\postingFilesWithoutStemming")).mkdir();
+                        Parse parse = new Parse(dirPath + "\\stop words");
+                        ReadFile readFile = new ReadFile(dirPath);
+                        Indexer indexer = new Indexer();
+                        Thread readFileThread = new Thread(readFile);
+                        Thread parseThread = new Thread(parse);
+                        Thread indexThread = new Thread(indexer);
+                        long startTime = System.nanoTime();
+                        startsIndexing = true;
+                        parseThread.start();
+                        readFileThread.start();
+                        indexThread.start();
+                        try {
+                            readFileThread.join();
+                            parseThread.join();
+                            indexThread.join();
+                        } catch (InterruptedException e) {
+                            e.printStackTrace();
+                        }
+                        // ------ THE FINAL ALERT BOX INDICATING THE INDEXING PROCESS IS DONE ------
+                        finally {
+                            startsIndexing = false;
+                            alreadyIndexed = true;
+                            double totalTimeInSeconds = (System.nanoTime() - startTime) * Math.pow(10, -9);
+                            int totalTimeInMinutes = (int) (totalTimeInSeconds / 60);
+                            int remainingSeconds = (int) (totalTimeInSeconds % 60);
+                            String totalTime = "Total time: " + totalTimeInMinutes + " minutes and " + remainingSeconds + " seconds.";
+                            String docCount = "Total documents indexed: " + ReadFile.docCount;
+                            String termCount = "Total unique words found: " + Indexer.totalUniqueTerms;
+                            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                            alert.setHeaderText("Indexing Done!");
+                            alert.setContentText(totalTime + "\n" + docCount + "\n" + termCount);
+                        }
+                    } else {
+                        Alert alert = new Alert(Alert.AlertType.ERROR);
+                        alert.setContentText("You must choose an existing stop words file path!");
+                        alert.show();
                     }
                 } else {
                     Alert alert = new Alert(Alert.AlertType.ERROR);
-                    alert.setContentText("You must choose an existing stop words file path!");
+                    alert.setContentText("You must choose an existing posting files path!");
                     alert.show();
                 }
-            }
-            else {
-                Alert alert = new Alert(Alert.AlertType.ERROR);
-                alert.setContentText("You must choose an existing posting files path!");
-                alert.show();
             }
         }
     }
@@ -116,17 +120,20 @@ public class Controller {
      * @param actionEvent - unused
      */
     public void onReset(ActionEvent actionEvent) {
-        corpusPath.setText("");
-        postingPath.setText("");
-        stemmingCheckBox.setSelected(true);
-        File dir = new File(postingPathText);
-        for (File dir2 : dir.listFiles()) {
-            for (File f : dir2.listFiles())
-                f.delete();
-            dir2.delete();
+        if (!startsIndexing) {
+            corpusPath.setText("");
+            postingPath.setText("");
+            stemmingCheckBox.setSelected(true);
+            File dir = new File(postingPathText);
+            for (File dir2 : dir.listFiles()) {
+                for (File f : dir2.listFiles())
+                    f.delete();
+                dir2.delete();
+            }
+            Indexer.termDictionary = null;
+            Indexer.documentDictionary = null;
+            alreadyIndexed = false;
         }
-        Indexer.termDictionary = null;
-        Indexer.documentDictionary = null;
     }
 
     /**
@@ -134,6 +141,22 @@ public class Controller {
      * @param actionEvent - unused
      */
     public void onDictionaryShow(ActionEvent actionEvent) {
+        if (!startsIndexing) {
+            try {
+                Stage stage = new Stage();
+                stage.setTitle("The Corpus's Dictionary");
+                FXMLLoader fxmlLoader = new FXMLLoader();
+                Parent root = fxmlLoader.load(getClass().getResource("/fxml/dictionary.fxml"));
+                fxmlLoader.setController(new DictionaryController(Indexer.getDictionaryString()));
+                Scene scene = new Scene(root, 600, 400);
+                stage.setScene(scene);
+                stage.initModality(Modality.APPLICATION_MODAL);
+                stage.show();
+            }
+            catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
     }
 
     /**
@@ -141,10 +164,22 @@ public class Controller {
      * @param actionEvent - unused
      */
     public void onDictionaryLoad(ActionEvent actionEvent) {
+        if (!startsIndexing) {
+            DirectoryChooser directoryChooser = new DirectoryChooser();
+            directoryChooser.setTitle("Choose a path");
+            File selectedDirectory = directoryChooser.showDialog(postingPath.getScene().getWindow());
+        }
+    }
+
+    /**
+     * chooses and sets a path for a given directory
+     * @param path - the text field of the given path
+     */
+    private void chooseAndSaveDirectoryPath(TextField path) {
         DirectoryChooser directoryChooser = new DirectoryChooser();
         directoryChooser.setTitle("Choose a path");
-        File selectedDirectory = directoryChooser.showDialog(postingPath.getScene().getWindow());
+        File selectedDirectory = directoryChooser.showDialog(path.getScene().getWindow());
         if (selectedDirectory != null)
-            postingPath.setText(selectedDirectory.getAbsolutePath());
+            path.setText(selectedDirectory.getAbsolutePath());
     }
 }
